+++ src/services/product-service.ts
@@ -6,7 +6,7 @@ class ProductService {
     async get () {
         const products = await prisma.product.findMany();
         if(!products) {
-            throw new BadRequestError("there is no products");   
+            throw new BadRequestError("There is no products");   
         }
         return products;
     }
@@ -18,7 +18,7 @@ class ProductService {
             },
         });
         if(!product) {
-            throw new NotFoundError("product", id);  
+            throw new NotFoundError(`Entity with id=${id} not found in product`);  
         }
         return product;
     }
@@ -33,25 +33,52 @@ class ProductService {
         return product;
     }
 
-    async update(name: string, id: number) {
+    async update(user_id: number, name: string, id: number) {
+        const _product = await prisma.product.findFirst({
+            where: {
+                id,
+                user_id,
+            },
+        });
+        if(!_product) {
+            throw new NotFoundError(`Entity with id=${id} not found in product`);
+        }
         const product = await prisma.product.update({
             where: {
-                id: Number(id),
+                id,
             },
             data: {
                 name: name,
             }
         });
-        if(!product) {
-            throw new BadRequestError();
-        }
         return product;
     }
 
-    async delete(id) {
+    async delete(id: number, user_id: number) {
+        const _product = await prisma.product.findFirst({
+            where: {
+                id,
+                user_id,
+            },
+        });
+        if(!_product) {
+            throw new NotFoundError(`Entity with id=${id} not found in product`);
+        }
+
+        const updates = await prisma.updatePoint.findMany();
+        if(updates) {
+            updates.forEach(async update => {
+                await prisma.updatePoint.delete({
+                    where: {
+                        id: update.id
+                    }
+                })
+            });
+        }
+
         const product = await prisma.product.delete({
             where: {
-                id: Number(id),
+                id,
             }
         })
         return product;
